requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
WTForms==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
alembic==1.13.2
gunicorn==22.0.0
pytz==2024.1
apscheduler==3.10.4
azure-identity==1.17.1
azure-mgmt-compute==33.0.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
email-validator==2.2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
email-validator==2.2.0
cryptography==43.0.0
passlib[argon2]==1.7.4
Flask-Limiter==3.5.1
limits==3.13.0
redis==5.0.8
